Move common/test to common
Resolves https://github.com/grafana/xk6-browser/pull/141#discussion_r764059420
