+++ common/connection_test.go
@@ -30,14 +30,13 @@ import (
 	"github.com/chromedp/cdproto/cdp"
 	"github.com/chromedp/cdproto/target"
 	"github.com/gorilla/websocket"
-	"github.com/grafana/xk6-browser/common/test"
 	"github.com/mailru/easyjson"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestConnection(t *testing.T) {
-	server := test.WSServerWithEcho(t)
+	server := WSServerWithEcho(t)
 	defer server.Cleanup()
 
 	t.Run("connect", func(t *testing.T) {
@@ -52,7 +51,7 @@ func TestConnection(t *testing.T) {
 }
 
 func TestConnectionClosureAbnormal(t *testing.T) {
-	server := test.WSServerWithClosureAbnormal(t)
+	server := WSServerWithClosureAbnormal(t)
 	defer server.Cleanup()
 
 	t.Run("closure abnormal", func(t *testing.T) {
@@ -70,7 +69,7 @@ func TestConnectionClosureAbnormal(t *testing.T) {
 }
 
 func TestConnectionSendRecv(t *testing.T) {
-	server := test.WSServerWithCDPHandler(t, test.CDPDefaultHandler, nil)
+	server := WSServerWithCDPHandler(t, CDPDefaultHandler, nil)
 	defer server.Cleanup()
 
 	t.Run("send command with empty reply", func(t *testing.T) {
@@ -134,7 +133,7 @@ func TestConnectionCreateSession(t *testing.T) {
 		}
 	}
 
-	server := test.WSServerWithCDPHandler(t, handler, &cmdsReceived)
+	server := WSServerWithCDPHandler(t, handler, &cmdsReceived)
 	defer server.Cleanup()
 
 	t.Run("create session for target", func(t *testing.T) {
