+++ common/session_test.go
@@ -31,7 +31,6 @@ import (
 	cdppage "github.com/chromedp/cdproto/page"
 	"github.com/chromedp/cdproto/target"
 	"github.com/gorilla/websocket"
-	"github.com/grafana/xk6-browser/common/test"
 	"github.com/mailru/easyjson"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -60,18 +59,18 @@ func TestSessionCreateSession(t *testing.T) {
 			case cdproto.MethodType(cdproto.CommandTargetAttachToTarget):
 				writeCh <- cdproto.Message{
 					Method: cdproto.EventTargetAttachedToTarget,
-					Params: easyjson.RawMessage([]byte(test.DEFAULT_CDP_TARGET_ATTACHED_TO_TARGET_MSG)),
+					Params: easyjson.RawMessage([]byte(DEFAULT_CDP_TARGET_ATTACHED_TO_TARGET_MSG)),
 				}
 				writeCh <- cdproto.Message{
 					ID:        msg.ID,
 					SessionID: msg.SessionID,
-					Result:    easyjson.RawMessage([]byte(test.DEFAULT_CDP_TARGET_ATTACH_TO_TARGET_RESPONSE)),
+					Result:    easyjson.RawMessage([]byte(DEFAULT_CDP_TARGET_ATTACH_TO_TARGET_RESPONSE)),
 				}
 			}
 		}
 	}
 
-	server := test.WSServerWithCDPHandler(t, handler, &cmdsReceived)
+	server := WSServerWithCDPHandler(t, handler, &cmdsReceived)
 	defer server.Cleanup()
 
 	t.Run("send and recv session commands", func(t *testing.T) {
@@ -82,9 +81,9 @@ func TestSessionCreateSession(t *testing.T) {
 
 		if assert.NoError(t, err) {
 			session, err := conn.createSession(&target.Info{
-				TargetID:         test.DEFAULT_CDP_TARGET_ID,
+				TargetID:         DEFAULT_CDP_TARGET_ID,
 				Type:             "page",
-				BrowserContextID: test.DEFAULT_CDP_BROWSER_CONTEXT_ID,
+				BrowserContextID: DEFAULT_CDP_BROWSER_CONTEXT_ID,
 			})
 
 			if assert.NoError(t, err) {
